Split functions on several files with headers
