+++ functions.h
@@ -1,0 +1,28 @@
+/*
+ * include/functions.h
+ *
+ * Some commands
+ *
+ */
+
+#ifndef _FUNCTIONS_H_
+#define _FUNCTIONS_H_
+
+int psh_cp(char **args);
+
+int psh_rm(char **args);
+
+int psh_touch(char **args);
+
+int psh_cat(char **args);
+
+int psh_man(char **args);
+
+int psh_cd(char **args);
+
+int psh_help(char **args);
+
+int psh_exit(char **args);
+
+
+#endif /* _FUNCTIONS_H_ */
